fix: resolve lint errors and format issues
- Fixed formatting in API types (gci alignment)
- Fixed deprecated comment casing (DEPRECATED -> Deprecated)
- Removed unused nolint directive
- Fixed badge controller int64 conversion
- Fixed unconvert errors in client code
- Fixed val assignments in project.go

Remaining: 1 typecheck error in auto-generated zz_project.go
(generator copies code from namespaced badge controller)

+++ pkg/cluster/clients/users/zz_runner.go
@@ -59,7 +59,7 @@ func GenerateGroupRunnerOptions(p *groupsv1alpha1.RunnerParameters) *gitlab.Crea
 	opts := generateCommonRunnerOptions(&p.CommonRunnerParameters)
 
 	if p.GroupID != nil {
-		groupID := int64(*p.GroupID)
+		groupID := *p.GroupID
 		opts.GroupID = &groupID
 	}
 
@@ -73,7 +73,7 @@ func GenerateProjectRunnerOptions(p *projectsv1alpha1.RunnerParameters) *gitlab.
 	opts := generateCommonRunnerOptions(&p.CommonRunnerParameters)
 
 	if p.ProjectID != nil {
-		projectID := int64(*p.ProjectID)
+		projectID := *p.ProjectID
 		opts.ProjectID = &projectID
 	}
 
@@ -94,7 +94,7 @@ func generateCommonRunnerOptions(p *commonv1alpha1.CommonRunnerParameters) *gitl
 		MaintenanceNote: p.MaintenanceNote,
 	}
 	if p.MaximumTimeout != nil {
-		maxTimeout := int64(*p.MaximumTimeout)
+		maxTimeout := *p.MaximumTimeout
 		opts.MaximumTimeout = &maxTimeout
 	}
 	return opts
